requirements.txt
Flask
langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf
faiss-cpu
sentence-transformers
transformers
accelerate
torch
bitsandbytes
langchain-huggingface
langchain-google-genai
google-generativeai
